Added GPGME
Also switched to cmake because holy heck is autotools slow and painful. I still don't know how to deal with the "plugin" style hooks loading up copies of all the symbols in each of their libraries. And I made the libevent2 base loop non-global, at the cost of destroying modularity and coupling libevent2 with f-ing everything.